1. MainIntent refactored 2. Screen handlers added 3. DailyRatesKeeper added 4. ExchangeScreenStateKeeper added 5. Minor improvements and refactorings

+++ app/src/main/java/org/convert/convercy/data/dto/CurrencyDescriptionDTO.kt
@@ -20,8 +20,8 @@ data class CurrencyDescriptionDTO(
     val name: String? = null,
 
     @field:SerializedName("Value")
-    val currency: Float? = null,
+    val currency: Double? = null,
 
     @field:SerializedName("Previous")
-    val previous: Float? = null
+    val previous: Double? = null
 )
